requirements.txt
requests==2.24.0
beautifulsoup4==4.9.3
google-api-python-client
google-auth-httplib2
google-auth-oauthlib